Use size_t instead of int

+++ Metric.h
@@ -1,6 +1,8 @@
 #ifndef _METRIC_H
 #define _METRIC_H
 
+#include <stddef.h>
+
 namespace SBody {
 	struct source {
 		const double mass;
@@ -11,13 +13,13 @@ namespace SBody {
 	namespace Metric {
 		// from cartesian to spherical
 		int c2s(const double x[], const double v[], double r[], double w[]);
-		int c2s(const double x[], double r[], const int dimension);
+		int c2s(const double x[], double r[], const size_t dimension);
 
 		// from spherical to cartesian
 		int s2c(const double r[], const double w[], double x[], double v[]);
-		int s2c(const double r[], double x[], const int dimension);
+		int s2c(const double r[], double x[], const size_t dimension);
 		namespace Newton {
-			extern const int dimension;
+			extern const size_t dimension;
 			extern int PN;
 			int function(double t, const double y[], double dydt[], void *params);
 			int jacobian(double t, const double y[], double *dfdy, double dfdt[], void *params);
@@ -25,7 +27,7 @@ namespace SBody {
 			double angularMomentum(const double x[], void *params);
 		} // namespace Newton
 		namespace Schwarzschild {
-			extern const int dimension;
+			extern const size_t dimension;
 			int function(double t, const double y[], double dydt[], void *params);
 			int jacobian(double t, const double y[], double *dfdy, double dfdt[], void *params);
 			double energy(const double y[], void *params);
@@ -34,7 +36,7 @@ namespace SBody {
 			int lightNormalization(double y[], void *params);
 		} // namespace Schwarzschild
 		namespace Kerr {
-			extern const int dimension;
+			extern const size_t dimension;
 			int function(double t, const double y[], double dydt[], void *params);
 			int jacobian(double t, const double y[], double *dfdy, double dfdt[], void *params);
 			double energy(const double y[], void *params);
@@ -44,7 +46,7 @@ namespace SBody {
 			int lightNormalization(double y[], void *params);
 		} // namespace Kerr
 		namespace KerrH {
-			extern const int dimension;
+			extern const size_t dimension;
 			int qdq2qp(const double r[], double u[], void *params);
 			int qp2qdq(const double u[], double r[], void *params);
 			int function(double t, const double y[], double dydt[], void *params);
